backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-cloud-storage==2.19.0
google-cloud-documentai>=2.29.0
google-auth==2.37.0
python-dotenv==1.0.1
PyMuPDF==1.25.3
pydantic==2.10.4
google-genai>=0.8.0
sqlalchemy>=2.0.0
google-generativeai>=0.8.0
numpy>=1.24.0
python-docx>=1.1.0
openpyxl>=3.1.0
